+++ src/main/java/plugily/projects/thebridge/handlers/language/LanguageManager.java
@@ -74,8 +74,17 @@ public class LanguageManager {
 
   private static void registerLocales() {
     Arrays.asList(
+      new Locale("Czech", "Český", "cs_CZ", "POEditor contributors", Arrays.asList("czech", "cesky", "český", "cs")),
+      new Locale("Dutch", "Nederlands", "nl_NL", "POEditor contributors", Arrays.asList("dutch", "nederlands", "nl")),
       new Locale("English", "English", "en_GB", "Tigerpanzer_02", Arrays.asList("default", "english", "en")),
-      new Locale("German", "Deutsch", "de_DE", "2Wild4You, Tigerpanzer_02 and POEditor contributors", Arrays.asList("deutsch", "german", "de")))
+      new Locale("French", "Français", "fr_FR", "POEditor contributors", Arrays.asList("french", "francais", "français", "fr")),
+      new Locale("German", "Deutsch", "de_DE", "2Wild4You, Tigerpanzer_02 and POEditor contributors", Arrays.asList("deutsch", "german", "de")),
+      new Locale("Hungarian", "Magyar", "hu_HU", "POEditor contributors", Arrays.asList("hungarian", "magyar", "hu")),
+      new Locale("Italian", "Italiano", "it_IT", "POEditor contributors", Arrays.asList("italian", "italiano", "it")),
+      new Locale("Polish", "Polski", "pl_PL", "Plajer", Arrays.asList("polish", "polski", "pl")),
+      new Locale("Portuguese (BR)", "Português Brasileiro", "pt_BR", "POEditor contributors", Arrays.asList("brazilian", "brasil", "brasileiro", "pt-br", "pt_br")),
+      new Locale("Russian", "Pусский", "ru_RU", "POEditor contributors", Arrays.asList("russian", "pусский", "pyccknn", "russkiy", "ru")),
+      new Locale("Spanish", "Español", "es_ES", "POEditor contributors", Arrays.asList("spanish", "espanol", "español", "es")))
       .forEach(LocaleRegistry::registerLocale);
   }