+++ src/main/java/plugily/projects/thebridge/handlers/setup/components/BaseComponent.java
@@ -40,6 +40,7 @@ import plugily.projects.thebridge.handlers.hologram.ArmorStandHologram;
 import plugily.projects.thebridge.handlers.setup.BaseUtilities;
 import plugily.projects.thebridge.handlers.setup.SetupInventory;
 import plugily.projects.thebridge.utils.CuboidSelector;
+import plugily.projects.thebridge.utils.Debugger;
 import plugily.projects.thebridge.utils.Utils;
 import plugily.projects.thebridge.utils.conversation.SimpleConversationBuilder;
 
@@ -85,6 +86,7 @@ public class BaseComponent implements SetupComponent {
           player.sendRawMessage(plugin.getChatManager().colorRawMessage("&e✔ Completed | &aColor of base " + getId(player) + " set to " + color));
           config.set("instances." + arena.getId() + ".bases." + getId(player) + ".color", color);
           ConfigUtils.saveConfig(plugin, config, "arenas");
+          BaseUtilities.addEditing(player);
           new SetupInventory(arena, player).openBases();
           return Prompt.END_OF_CONVERSATION;
         }
@@ -108,6 +110,7 @@ public class BaseComponent implements SetupComponent {
       LocationSerializer.saveLoc(plugin, config, "arenas", "instances." + arena.getId() + ".bases." + getId(player) + ".baselocation2", selection.getSecondPos());
       arena.setEndLocation(player.getLocation());
       player.sendMessage(plugin.getChatManager().colorRawMessage("&e✔ Completed | &aBase location for arena " + arena.getId() + " set with your selection!"));
+      BaseUtilities.addEditing(player);
       ConfigUtils.saveConfig(plugin, config, "arenas");
     }), 1, 0);
 
@@ -130,12 +133,13 @@ public class BaseComponent implements SetupComponent {
 
       player.sendMessage(plugin.getChatManager().colorRawMessage("&e✔ Completed | &aPortal location for arena " + arena.getId() + " set with your selection!"));
       player.sendMessage(plugin.getChatManager().colorRawMessage("&e✔ Completed &cautomatically &e| &aPortalHologram location for base " + getId(player) + " set at the mid of your selection! Feel free to change it if you want!"));
+      BaseUtilities.addEditing(player);
       ConfigUtils.saveConfig(plugin, config, "arenas");
     }), 2, 0);
 
     pane.addItem(new GuiItem(new ItemBuilder(XMaterial.ENDER_EYE.parseMaterial())
-      .name(plugin.getChatManager().colorRawMessage("&e&lSet Cage Location"))
-      .lore(ChatColor.GRAY + "Click to set the cage location")
+      .name(plugin.getChatManager().colorRawMessage("&e&lSet Cage Location (Only floor)"))
+      .lore(ChatColor.GRAY + "Click to set the cage location (only floor needed)")
       .lore(ChatColor.GRAY + "after you selected it with the location wand")
       .lore(ChatColor.DARK_GRAY + "(Please just select the blocks that should be removed/set)")
       .lore("", setupInventory.getSetupUtilities().isOptionDoneBool("instances." + arena.getId() + ".bases." + getId(player) + ".cagelocation1"))
@@ -146,10 +150,15 @@ public class BaseComponent implements SetupComponent {
         player.sendMessage(plugin.getChatManager().colorRawMessage(plugin.getChatManager().getPrefix() + "&cPlease select both corners before adding an base location!"));
         return;
       }
+      if (new Cuboid(selection.getFirstPos(), selection.getSecondPos()).contains(XMaterial.AIR.parseMaterial())) {
+        player.sendMessage(plugin.getChatManager().colorRawMessage(plugin.getChatManager().getPrefix() + "&cPlease select only the floor of the cage! Make sure that it is not air!"));
+        return;
+      }
       LocationSerializer.saveLoc(plugin, config, "arenas", "instances." + arena.getId() + ".bases." + getId(player) + ".cagelocation1", selection.getFirstPos());
       LocationSerializer.saveLoc(plugin, config, "arenas", "instances." + arena.getId() + ".bases." + getId(player) + ".cagelocation2", selection.getSecondPos());
 
       player.sendMessage(plugin.getChatManager().colorRawMessage("&e✔ Completed | &aCage location for arena " + arena.getId() + " set with your selection!"));
+      BaseUtilities.addEditing(player);
       ConfigUtils.saveConfig(plugin, config, "arenas");
     }), 3, 0);
 
@@ -166,6 +175,7 @@ public class BaseComponent implements SetupComponent {
       e.getWhoClicked().closeInventory();
       config.set("instances." + arena.getId() + ".bases." + getId(player) + ".spawnpoint", serializedLocation);
       player.sendMessage(plugin.getChatManager().colorRawMessage("&e✔ Completed | &aSpawnPoint location for base " + getId(player) + " set at your location!"));
+      BaseUtilities.addEditing(player);
       ConfigUtils.saveConfig(plugin, config, "arenas");
     }), 4, 0);
 
@@ -180,6 +190,7 @@ public class BaseComponent implements SetupComponent {
       e.getWhoClicked().closeInventory();
       config.set("instances." + arena.getId() + ".bases." + getId(player) + ".respawnpoint", serializedLocation);
       player.sendMessage(plugin.getChatManager().colorRawMessage("&e✔ Completed | &aReSpawnPoint location for base " + getId(player) + " set at your location!"));
+      BaseUtilities.addEditing(player);
       ConfigUtils.saveConfig(plugin, config, "arenas");
     }), 5, 0);
 
@@ -195,6 +206,7 @@ public class BaseComponent implements SetupComponent {
         player.sendMessage(plugin.getChatManager().colorRawMessage("&cLocation changes take affect after restart!"));
       }
       player.sendMessage(plugin.getChatManager().colorRawMessage("&e✔ Completed | &aPortalHologram location for base " + getId(player) + " set at your location!"));
+      BaseUtilities.addEditing(player);
       ConfigUtils.saveConfig(plugin, config, "arenas");
     }), 6, 0);
 
@@ -249,6 +261,7 @@ public class BaseComponent implements SetupComponent {
       base.setArmorStandHologram(portal);
       ConfigUtils.saveConfig(plugin, config, "arenas");
       BaseUtilities.getBaseId().remove(player);
+      BaseUtilities.removeEditing(player);
     }), 7, 0);
   }