+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### 1.0.6 Release (13.01.2021 - 18.01.2021)
+* ArenaRegistry will set isDone to false if arena cuboid is not big enough
+* Fixed joining through a sign while hotbar slot for leave-item is active
+* Fixed Cage doesn't disappear on first round
+* Added error message if cage floor contains Air
+* Added customizable item name in arena selector gui (by ajgeiss0702)
+* Added preconfigured locales for Czech, Dutch, French, Hungarian, Italian,
+  Polish, Portuguese, Russian, Spanish [Thanks to PoEditor contributors]
+* Changed SetupInventory to open base menu while editing base
+* Fix wrong teleport location on arena leave
+
 ### 1.0.5 Release (05.01.2021 - 12.01.2021)
 * Fixed Teleportation to the end is now no longer possible
 * Fixed NoSuchMethod isAir check